Fixed issue to properly create the config file
I noticed that the config file has been omitted from the tag branches.  When we require

"bllim/datatables": "*"

from composer, it grabs the latest 1.3.3 tag version, but this is missing that important config file.

There are two ways to fix this, to require the master branch in composer which does contain the file, or make a new tag with the latest files from master and confirming that it contains the /config file.

I have also updated the php artisan command, since the one currently listed there is not the correct one.